Fix script behaviour on MacOS
On MacOS, gtkwave must be run with the system `open` command (see the
note in the README for a link to a resource explaining the details).
Unfortunately, this prevents piping stdout, so instead gtkwave must be
run with the `-O` flag to direct its output to a file. However, this
captures all of gtkwaves logging output as well, corrupting the GTKW
file we are writing to. We use awk along with start and end markers to
account for this.

This commit also adds functionality in that we now ignore any signals
whose name begins with `_`. This plays nicely with Chisel where all
the temporary signals generated begin with `_`. In a large design the
temporary signals can greatly outnumber that actual signals of interest,
so ignoring them is crucial.